♻️ : DataUtils 분리
