+++ src/Utils/ValidationUtils.ts
@@ -1,0 +1,7 @@
+export const isNumberInRange = (value: number, min: number, max: number): boolean =>
+  value >= min && value <= max;
+
+export const isStringLengthInRange = (value: string, min: number, max: number): boolean =>
+  value.length >= min && value.length <= max;
+
+export const isCorrectNumberUnit = (value: number, unit: number): boolean => value % unit === 0;
