+++ src/Utils/index.ts
@@ -1,6 +1,6 @@
-export * from 'Utils/DataUtils';
+export * from 'Utils/ParserUtils';
 export * from 'Utils/EventUtils';
 export * from 'Utils/ElementUtils';
-export * from 'Utils/ElementUtils';
+export * from 'Utils/ValidationUtils';
 export * from 'Utils/User/validator';
 export * from 'Utils/snackbar';
